New DSN parser (#400)
This is a rewrite of the gel-tokio DSN parser to make it more testable, and to not require an external environment to do so.

The parser allows for a mock environment, filesystem and user profile setup to better test without requiring live filesystem modifications.

The API of the Builder changed slightly, but efforts were made to maintain the older API. The build_env function was deprecated.

Note that unnecessary async was removed from the builder. The likelihood of file access using this this DSN builder causing major stalls is very low. We can use classic blocking operations.

This continues to be the easiest way to create a config:

let config = Builder::default()::build().unwrap();

We now add intentional ways to configure a client without the environment or filesystem (mainly for testing purposes):

let config = Builder::default().host(...).port(...).without_system().build().unwrap();