Add stored credentials API (#431)
We currently have N paths which read and write the credentials. This creates a centralized API in gel-dsn that the CLI can use to read and write them as well.

In addition, we now remove any default branch/database settings in credential files to help prevent issues when moving between CLI versions.