Rework protocol generation to clean up support enums (#453)
This is a rework of the protocol generation crate to reduce some of the macro magic in favour of more traits. While still somewhat complex, it's a little easier to work with this iteration of the protocol generator.

We also now support enums (non-bitfield-only for now).

Both the PostgreSQL and EdgeDB protocols have been fully updated, though neither of them is in active use (yet). The EdgeDB protocol in gel-protocol will be replaced with this new implementation once it can be confirmed that it behaves as expected.

Measurement structs were eliminated because they turned out not to be useful. We rarely measure data, and builders are good enough for that. The Meta types were also eliminated in favour of a const StructFieldMeta which works in a somewhat similar way but eliminated a number of types in favour of just one.