Ignore missing CLOSE_NOTIFY (#461)
On Windows, if a FIN is received on a socket before CLOSE_NOTIFY has been processed, the Rustls client may believe that the server failed to send it when in fact, it was dropped by WinSock because EOF on a TCP stream throws away the read buffer.

We add an option to ignore the missing close_notify and thread it through the gel-stream library. This is only enable on Windows.