Implement previewable streams (#466)
This implements preview-able streams for server-side OpenSSL and Rustls connections. A preview-able stream allows you to peek at the first N bytes of a stream (8 is the default), to determine which protocol you are about to serve.

This can be used to handle raw or TLS-wrapped protocols of various types and dispatch based on those first N bytes.

The Acceptor without previewing is unchanged: it returns a stream of connections. The Acceptor with previewing returns a stream of (Preview, Connection) instead, allowing you to peek at the bytes on the connection and potentially call secure_upgrade (or secure_upgrade_preview) on that connection.