Fix and update tests for SNI-based TLS lookup (#467)
We previously had a with_lookup for TlsServerParameterProvider, but it was somewhat limited and we didn't actually have any specific tests for it.

This is now implemented, as we test that we can switch certificates and ALPN based on the SNI sent in the ClientHello.

Note that we also now offer the stream metadata as part of the TLS parameter lookup: this can allow an embedder to also modify certs and/or add/remove ALPN entries depending on the underlying stream properties (remote address, transport).

To simplify the internal implementation, the secure_upgrade method now consumes the stream and returns it when the upgrade completes.

This also requires a Rust MSRV bump so we get access to the trait upcasting feature.