Multi-protocol gel frontend crate (#457)
This is the initial implementation of "babelfish", a network frontend that speaks:

    Gel/EdgeDB
    PostgreSQL
    HTTP 1/2
    SSL
    ... and potentially more

This crate deals with connection termination and credential verification. Authorization/credential management and actual stream handling are designed to be punted to other layers of the system.

The code is not complete, but we are able to speak enough that the psql and gel CLIs both get responses they understand, and HTTP/HTTPS is working. All of these work over both TLS and plaintext.