Use the new gel-db-protocol to decode all client structs (#498)
As part of the protocol message validation, we want to use the new protocol to decode all client message structs to ensure we're speaking the protocol properly.

This is a mostly-mechanical translation of the decoding functionality to use new_protocol. There are a lot of copies here because we don't want to make too many changes at once.

This code will be somewhat awkward until we finish migrating encoding and decoding.