Move gel-config from gel-cli (#522)
This is mostly a straight port from gel-cli. Once we've imported this code and updated gel-cli, I'm going to take a pass at making it more generic to support both CLI and server.

See https://github.com/geldata/gel-cli/pull/1721 for supporting CLI changes (these changes pass locally for me with an overridden cargo config).

Some minor changes:

 - anyhow replaced with strict error enum type
 - primitive types are now an enum rather than strings

A pseudo-debug syntax was added to the commands so that we could more easily test the output of the parse results, but this is not currently checked by anything other than eyeball.